+++ tools/midiChecker.ts
@@ -1,6 +1,7 @@
 
 import fs from "fs";
 import {argv} from "yargs";
+import YAML from "yaml";
 import {MIDI, MusicNotation, Matcher} from "@k-l-lambda/web-widgets";
 
 import "../env.js";
@@ -52,6 +53,12 @@ const matchMIDI = async (criterion: MIDI.MidiData, sample: MIDI.MidiData): Promi
 const main = async () => {
 	const lilyParser = await loadLilyParser();
 
+	const logStream = fs.createWriteStream("midiChecker.log");
+	const log = (...messages) => {
+		console.log(...messages);
+		logStream.write(messages.join(" ") + "\n", "utf-8");
+	};
+
 	if (argv.inputLy) {
 		const source = fs.readFileSync(argv.inputLy).toString();
 		const midi = await ScoreMaker.makeMIDI(source, lilyParser);
@@ -70,33 +77,64 @@ const main = async () => {
 		const counting = {
 			count: 0,
 			perfect: 0,
+			failure: 0,
 			coverageSum: 0,
 		};
+
+		const issues = [];
 
 		for (const lyFile of lyFiles) {
 			const midiFile = lyFile.replace(/\.\w+$/, ".mid");
 			if (fs.existsSync(midiFile)) {
 				//console.log("pair:", lyFile, midiFile);
-				const filename = lyFile.match(/[^/]+$/)[0];
+				try {
+					const filename = lyFile.match(/[^/]+$/)[0];
 
-				const ly = fs.readFileSync(lyFile).toString();
-				const midiBuffer = fs.readFileSync(midiFile);
+					const ly = fs.readFileSync(lyFile).toString();
+					const midiBuffer = fs.readFileSync(midiFile);
+	
+					const criterion = MIDI.parseMidiData(midiBuffer);
+					const sample = await ScoreMaker.makeMIDI(ly, lilyParser);
+	
+					const result = await matchMIDI(criterion, sample);
+	
+					if (result.coverage === 1)
+						++counting.perfect;
+					else {
+						issues.push({
+							path: lyFile,
+							coverage: result.coverage,
+							omitC: result.omitC,
+							omitS: result.omitS,
+						});
+					}
+					counting.coverageSum += result.coverage;
+	
+					console.log("match:", result.coverage.toFixed(3), [result.omitC, result.omitS], filename);
+				}
+				catch (err) {
+					++counting.failure;
 
-				const criterion = MIDI.parseMidiData(midiBuffer);
-				const sample = await ScoreMaker.makeMIDI(ly, lilyParser);
-
-				const result = await matchMIDI(criterion, sample);
+					issues.push({
+						path: lyFile,
+						coverage: -1,
+					});
+				}
 
 				++counting.count;
-				if (result.coverage === 1)
-					++counting.perfect;
-				counting.coverageSum += result.coverage;
-
-				console.log("match:", result.coverage.toFixed(3), [result.omitC, result.omitS], filename);
 			}
 		}
 
-		console.log("count:", counting.count, "perfect:", counting.perfect, "coverage:", counting.coverageSum / counting.count);
+		issues.sort((i1, i2) => i1.coverage - i2.coverage);
+
+		if (issues.length) {
+			log("Issues:");
+			log(YAML.stringify(issues));
+		}
+		else
+			log("No issues.");
+
+		log("MIDI checking finished, count:", counting.count, "perfect:", counting.perfect, "failure:", counting.failure, "coverage:", counting.coverageSum / counting.count);
 	}
 
 	console.log("Done.");